Update patch set 4
Patch Set 4:

According to the doc, I thought we only needed sha256 once we're using TPM 2.0, so I figured the way that gbb_utility is invoked would be a configuration option somewhere in the factory flow. 

However, both of these are really overkill. The whole point as I understand it is to provide some userspace app a way to say "Yes, I'm sure this is product FOO, not product BAR, so I can safely enable/disable feature SPLAT". But the basis of that attestation is the assumption that no one has disabled write-protect, changed the HWID, and re-enabled write-protect (or just replaced the firmware with one that lies, although that's more technically challenging). 

The same assertion and level of trust could be obtained by simply running "crossystem wpsw_boot fwid" and comparing the output with an expected string. But hey, whatever, "Trusted Computing" and all that. Me b'long cargo.

Patch-set: 4