Update patch set 4
Patch Set 4:

It's not the physical hardware mod threat that this protects against. Consider the case of a system compromise when the attacker does not have physical access. A local software utility is not reliable in this case but a quoted PCR that was extended by RO firmware is.

Patch-set: 4