Update patch set 4
Patch Set 4: -Commit-Queue

Not marking this ready until we've heard from David. If it's okay to use the first 20 bytes of the sha256 to extend the PCR on TPM 1.x (which seems to allow only 20 bytes for the TPM_Extend operation), then I'll need to tweak this a bit.

Patch-set: 4