vboot2: add bare hash algorithms
This adds signing algorithms for bare SHA1, SHA-256, and SHA-512.
This enables future vboot2 changes to treat signatures and hashes more
equivalently, which will be convenient for multiply-signed keyblocks.

Note that the key is still what determines which algorithm should be
used, not the signature.  Since we only generate keys with RSA+SHA
algorithms at present, this new code is used only in the unit tests.

BUG=chromium:423882
BRANCH=none
TEST=VBOOT2=1 make runtests

Change-Id: Ieeb19add1083e6382a33eac2ddc0b2d40a872efc
Signed-off-by: Randall Spangler <rspangler@chromium.org>