Update patch set 1
Patch Set 1: Code-Review-1

I'm not sure RO_NORMAL is the right flag.  We only used that with u-boot, so I think that flag is absent all the the time now (and vboot2 doesn't even support it).

So, two cases:

1) VbInit() called after option rom init.  OPROM_MATTERS applies, and we must reboot if OPROM_LOADED doesn't match the nvram flag.

2) VbInit() called before option rom init.  Don't set OPROM_MATTERS.  Initialize vbios the first time a screen is requested by vboot.

And maybe one third corner case:

3) VbInit() called before option rom init.  But kernel code will be called after option rom init.  So VbInit() needs to propagate the nvram flag to an output flag, but doesn't need to request a reboot if the nvram flag mismatches OPROM_LOADED.

And for (3), we need to handle that in both the is_dev case (to request oprom for the dev screen) and in the normal boot case (to request oprom for the slow-firmware-update screen).

Patch-set: 1
Reviewer: Gerrit User 1000187 <1000187@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1