Update patch set 1
Patch Set 1:

Yeah I chose RO_NORMAL exactly because it was what we used with U-boot when we did not call VbInit in firmware.

I can't not set OPROM_MATTERS because so much of the logic in vboot is tied in with oprom matters so it would be a much larger change.

Basically though, I need this in the samus firmware branch now to unblock a firmware update for the EVT2 build and I don't have a unit with me or the time to try a lot of more complicated options.

I can leave this out for now on TOT as other things are debated.

Patch-set: 1