RFC: Support writing GPT structs to off device location
This CL allows the GPT headers and partition entry arrays to be stored
in a different device. Instead of treating both the SPI and NAND devices
as one (in a sandwich way), this CL writes and reads the GPT struct
independently of the actual device that houses the partitions.
Therefore, the first usable LBA of the partitions will be at 0, and the
last usable LBA is at the end of the NAND.

  +------------------------+
  | SPI houses GPT structs |
  +------------------------+
    |
  0 |  Index into
  v v
  +------------------------+
  | NAND houses partitions |
  +------------------------+

Note that the "my_lba" and "alternate_lba" in the GPT headers are bogus.

BUG=chromium:425677
BRANCH=none
TEST=unittest
TEST=cgpt create/add/show on a stumpy-moblab

Change-Id: I083b3c94da3b0bb3da1a7b10c6969774080a2afd