vboot: Plumb the two disk sizes and 'gpt on device' param through
To support an external GPT, disks have two new attributes:
- A binary flag indicating whether the GPT is in the same address
  space as the payloads or a separate one.
- The number of sectors of the streaming portion of storage, as
  opposed to the portion containing the GPT.
These have been added elsewhere to GptData (in cgptlib) and BlockDev
(in depthcharge). This patch adds the plumbing between those, including
in the DiskInfo interface between the firmware and vboot.

A previous version of this patch had a bug, that it didn't handle
existing callers of vboot which don't set stream_lba_count. I had
made a mistake by mailing a version which didn't handle this after
testing a version which does handle this case. The fix is on
lines 106-108 of firmware/lib/vboot_api_kernel.c.

BUG=chromium:425677
BRANCH=none
TEST=Interactively wrote the GPT with cgpt and observed the following
boot with depthcharge to read the GPT from SPI and then read from
the proper locations in NAND flash.
make runalltests passes.
Storm boots from USB with the current ToT depthcharge (wasn't true
in the earlier revision.)

Signed-off-by: Dan Ehrenberg <dehrenberg@chromium.org>
Change-Id: Iacc4c97652ea16263a0e7ff09559b4889211a9a9