Update patch set 5
Patch Set 5: Code-Review+1

I have to say, I think your previous phrasing, with ShouldStoreOffDevice as a separate function, was cleaner than what you just changed it to. I was just saying we should exit the program if stat fails. But this is a minor style point and I'm not going to hold back the review for it.

I'm giving you a +1 and I'll let Bill (or Randall) do the final review, since this is not my code base.

Patch-set: 5
Label: Code-Review=+1