Update patch set 6
Patch Set 6:

Based on a thread with the factory team, it looks like at least through EVT, we'll be doing the initial installation as a netboot or something else running on the device. How about we let this version which only works in that mode go in, and the version which works on "copy machines" goes in later?

I agree it wouldn't be a bad idea to make the path to flashrom be a command-line parameter. (Even then, I think it's still good to use stat() to determine whether it's an MTD device so that it can complain early about not being passed that parameter/being passed it inappropriately.) Also it'd be nice if we had some library to construct the shell commands properly sanitized rather than using snprintf. Bonus points to applying the same treatments to vpd!

Patch-set: 6