Update patch set 7
Patch Set 7: Code-Review+1

This looks good to me. In a future CL, rather than just parametrizing flashrom, how about you do the full thing of removing the call to flashrom from here and putting that in a wrapper script, as called for in https://code.google.com/p/chromium/issues/detail?id=432611 ? And for now, how about we put the code in as is, just to allow firmware and kernel development to continue, while the cleaner fix to that bug is developed in parallel?

I'll leave this for Bill to +2.

Patch-set: 7
Label: Code-Review=+1