Update patch set 4
Patch Set 4: Code-Review+2 Commit-Queue+1

Addressed Bill's comments.  These were relatively minor, and he'd already +1'd, so I'm going to +2 this since it's blocking 5 other CLs that he's +2'd on the assumption that he just missed this one.  This is a code path which is not currently used in firmware, so if I've misjudged, I can address any other issues quickly and with no impact on any current board.

Patch-set: 4
Label: Code-Review=+2
Label: Commit-Queue=+1