Update patch set 2
Patch Set 2:

(5 comments)

Many thoughts.  I'd also like to get Bill's input.  Given how hard it is to change the signing process, we shouldn't rush this change in.  Any changes we put in we'll be living with for a long time.

The trickiest bit is how we decide whether to allow hw crypto or not.  If we're willing to allow it unconditionally, this becomes pretty simple (just call the new vb2ex_hwcrypto apis in 2sha_utility.c).

Patch-set: 2