Add cgpt_wrapper binary to transparently support NAND
This CL adds a "cgpt_wrapper" binary that will forward to "cgpt" binary
as needed to transparently support NAND. The corresponding ebuild will
need to install "cgpt" as "cgpt.bin" and this "cgpt_wrapper" as "cgpt".

The idea is that if we are working with an MTD device, this binary will
exec out to "flashrom" to read in the GPT section, then exec out to the
actual "cgpt" binary to work on that file, and finally write it back
with "flashrom".

BUG=chromium:442518
BRANCH=none
TEST=unittest and some trial runs with mtdram

Change-Id: I54534afa9a970ec858f313f698c0eb210c827477