Update patch set 2
Patch Set 2:

(3 comments)

I expected this to be written in shell script. Why did you decide to write it in C? I think it'll be easier to read and write this code in shell, and shell makes it trivial to escape arguments to binaries that are called.

This will need to be tested on real hardware before it is submitted. Let me know if you need help setting up the Storm board I gave you. What sort of testing did you do with mtdram? Did that somehow stand in as the target to flashrom?

Patch-set: 2
Label: Code-Review=0