futility: Support both vboot2 versions at once (2.0 and 2.1)
This CL touches a lot of files, but isn't as bad as it appears.

We had been trying to keep the changes between firmware/lib20 and
firmware/lib21 as small as possible, so we were using the exact
same names for all files and functions and only compiling with
one or the other.

This renames everything in both of those directories so that all
the exported functions have different names. Structures and
functions that are identical are left in firmware/2lib/ where
they were before, with only minor exceptions.

Because we can link with everything at once, we can now create
one large UTILLIB with all the needed components. When building
for the target we should only support one API. It's pretty clean
now, but making it completely clean will require a future CL.

BUG=chromium:228932
BRANCH=none
CQ-DEPEND=CL:235950
TEST=manual

  cros_workon --host start vboot_reference
  sudo emerge vboot_reference

  cros_workon-samus start vboot_reference coreboot
  emerge-samus vboot_reference coreboot

  cros_workon-veyron_pinky start vboot_reference coreboot
  emerge-veyron_pinky vboot_reference coreboot

  \rm -rf build; make runtests
  \rm -rf build; make runtests MINIMAL=1 VBOOT1=  VBOOT20=  VBOOT21=
  \rm -rf build; make runtests MINIMAL=1 VBOOT1=  VBOOT20=  VBOOT21=1
  \rm -rf build; make runtests MINIMAL=1 VBOOT1=  VBOOT20=1 VBOOT21=
  \rm -rf build; make runtests MINIMAL=1 VBOOT1=  VBOOT20=1 VBOOT21=1
  \rm -rf build; make runtests MINIMAL=1 VBOOT1=1 VBOOT20=1 VBOOT21=
  \rm -rf build; make runtests MINIMAL=1 VBOOT1=1 VBOOT20=  VBOOT21=1
  \rm -rf build; make runtests MINIMAL=1 VBOOT1=1 VBOOT20=1 VBOOT21=
  \rm -rf build; make runtests MINIMAL=1 VBOOT1=1 VBOOT20=1 VBOOT21=1

Change-Id: Id47f365c2d08bee03ea11b1267c43c725d7bf218
Signed-off-by: Bill Richardson <wfrichar@chromium.org>