sign_official_build.sh should work without dev-firmware keys
The signing scripts pass a bunch of args around, including paths
to the keys used to sign dev-mode-specific firmware. That was
only used on Alex and ZGB, so all the newer systems don't have
those keys and the script falls over.

This uses the normal firmware keys if the dev-firmware keys don't
exist. This was an oversight with the original CL that touched
resign_firmwarefd.sh

BUG=chromium:453901
BRANCH=ToT
TEST=manual

Download a newer signed recovery image, say for nyan, and save it as

  scripts/image_signing/nyan-recovery-mp.bin

Temporarily delete the developer firmware keys from the devkeys:

  rm -f tests/devkeys/dev_firmware*

Now try resigning the recovery image:

  cd scripts/image_signing
  ./sign_official_build.sh recovery nyan-recovery-mp.bin \
                           ../../tests/devkeys/ signed.bin \
                           ../../tests/devkeys/key.versions

It should work.

Change-Id: I474811158cb33e16ad09c16b0db825c40217dd70
Signed-off-by: Bill Richardson <wfrichar@chromium.org>