Update patch set 2
Patch Set 2: Verified+1 Trybot-Ready+1

Ok, the /bin/cgpt from 6699 failed to find cgpt.bin, but CL:242294 landed after that. When it failed, it was just returning "-1" with no output, making it a little more hard to diagnose what was going on.

We do call "cgpt" without the absolute path in scripts like the recovery script. Indeed, cgpt is installed in /bin in the initramfs but /usr/bin in the rootfs; so I'd recommend calling just "cgpt".

I changed this CL to print error when execv() fails and return a different error code.

Patch-set: 2
Label: Trybot-Ready=+1
Label: Verified=+1