vb21: Replace the key GUID with a sha1sum instead
We wanted a quick and human-friendly way to match keys with
signatures, so we decided to give each keypair a unique GUID and
carry that ID around when signing things.

But then we realized that we could autogenerate a unique
identifier from the .pem file itself, which is even better
because then we can match our keys with the openssl file used to
generate them.

This change replaces the GUID id with a sha1sum calculated from
the public key's "keyb" blob.

BUG=none
BRANCH=none
TEST=make runtests

Also:

  futility show tests/testkeys/key_rsa4096.pem
  futility create tests/testkeys/key_rsa4096.pem foo
  futility show foo.vbp*

Note that the GUID is the same for all files.

Change-Id: If1e624550c7ae4a06195a0d8fd55b6660c0dccbc
Signed-off-by: Bill Richardson <wfrichar@chromium.org>