Update patch set 2
Patch Set 2:

> Note if people are doing --force (as the original thread
 > mentioned), they'll bypass all checks and fall into same error.

True but we print a pretty big banner warning them that they are on their own if they do that.  I don't think it's worth trying to protect someone who thinks it's a good idea to use --force when they don't know exactly what the command is going to do.

Patch-set: 2