Update patch set 1
Patch Set 1:

Why do you want to do this for a rambi?  It's a laptop, right?  So it has a trusted keyboard.

We can switch to this but then you also have to change the screen displayed by the firmware, and in addition we become inconsistent with other models.

Patch-set: 1