enable USB boot on transition to dev on some devices
Some Chrome OS devices do not allow to login even in developer mode,
as they do not have display/keyboard and sshd is not part of the
Chrome OS imavge. Even enabling developer mode on those devices is
very involved (requires taking the device apart and is guaranteed to
take long time).

We still want to allow the end user to control those devices in dev
mode. The solution is enabling the ability to boot from the USB stick
when the device transitions from normal to developer mode.

A simple way to do it is to set the NVRAM flag, which allows USB boot.
The flag is set on normal=>dev transition only, and only on those
devices where it is configured (as discovered by invoking
VbExGetSwitches with the appropriate parameters).

BRANCH=storm
BUG=chrome-os-partner:38303
TEST=tested with the corresponding depthcharge patches

Change-Id: I5fa58963256598cde3b534f5250101fba6042f8c
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>