Report if firmware is signed by developer key
Recent experience shows that users often get confused and try running
pre-mp signed images under dev firmware control and vice versa. The
matters are further aggravated by the fact that the signage mismatch
is allowed when the device is in dev mode and not in normal mode.

While the users usually can tell what class of keys the Chrome OS
image is signed with, it is much mode difficult to tell what keys the
firmware was signed with.

This patch, reports in the log if the firmware was signed with dev
keys, by comparing the hash calculated over the key modulo and r^2
fields with a precompiled value.

BRANCH=none
BUG=none
TEST=booted the new code on storm, observed the following message
     included in the log:

  VB2:vb2_report_key_class() This is developer signed firmware

Change-Id: I97ed6ba384cee59ff3f42943630e92ebae10dd03
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>