Update patch set 4
Patch Set 4: Verified+1

ok, after offline chat with Randall I tried to verify the root key against its precompiled digest, using firmware/lib/vboot_display.c:FillInSha1Sum() as an example.

The thing is, however, that the suggested example uses SHA1 digest, which is marked as obsolete in vboot2, and it calculates the digest over a different key structure (vboot1 type).

Having looked at both variants (patch1 and patch2) I think patch1 is a bit easier to maintain - if the key is ever changed, it is easier to make the function work again. In case of patch2 one would need to calculate the digest of the new key and then edit the source.

But either way is better than nothing - I fielded a lot of third partly user complaints in the past month because they couldn't tell what kind of firmware they're running.

Patch-set: 4
Label: Verified=+1