Fix up unintention division by a 64-bit value
This is quite inefficient and is not supported in U-Boot or Linux. It is
also almost never needed. Adjust the sites that use this to avoid it.

For the one file where this is not trivial, provide a do_div() function
which divides by a 32-bit value and returns the 32-bit remainder.

BUG=chromium:488858
TEST=with other changes, can build a boot image with
  USE="-depthcharge" emerge-nyan_big chromeos-bootimage
BRANCH=none

Change-Id: I4f36262a2985c7d59cb42ffdf9a8d871e18265b0