vboot_reference: snarf local copy of /usr/include/tss/tcs.h
This goes against my instincts, but tcs.h only reflects the TPM 1.2
spec so it's pretty much guaranteed not to change.  We already
sprinkle a number of TPM/TCS/TSS constants in various places,
making the same assumption.

This is necessary to break a circular DEPENDency as we want to
send UMA stats from tcsd.  Without this, metrics depends on
vboot_reference which depends on trousers which depends on
metrics.  Technically the vboot_reference dependency on trousers
is header-file only, but we can't cope with that.

BUG=chromium:481552
TEST=compiled with emerge-<something> vboot_reference

Change-Id: Icb0e0b62c52dfa490050c533da4b4380ddb837d7