Update patch set 1
Patch Set 1: Code-Review+2

One more weird issue is that this code is still (I think?) only compiled into the fwlib20/fwlib21 targets, but cbfstool is a host utility. I assume we can still make it work that way (or are there any weird compiler environment settings for fwlib that would cause problems when linked into host tools?) but it would make more sense if it was part of libvboot_host.a, libvboot_util(21)?.a, or maybe a new libvboot_crypto.a. Guess we'll deal with that in a later refactoring?

Patch-set: 1
Label: Code-Review=+2