Update patch set 1
Patch Set 1:

This is for reboot-bootloader. OS needs to set the recovery reason field to some non-zero value to trigger into fastboot recovery mode. Any non-zero value would work here, but I think it would be a good idea to have an explicit reason value that can be used for this purpose.

Patch-set: 1